fix the following warning in r6194 - add part of work for #142 - generic send queue for backchannel is necessary in gfmd: {{{ netsendq.c:244: warning: format '%d' expects type 'int', but argument 8 has type 'long unsigned int' netsendq.c:250: warning: format '%d' expects type 'int', but argument 9 has type 'long unsigned int' }}}
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@6195 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/netsendq.c
@@ -242,13 +242,13 @@ netsendq_new(struct netsendq_manager *manager, struct abstract_host *abhost,
 	GFARM_MALLOC(qhost);
 	if (qhost == NULL) {
 		gflog_debug(GFARM_MSG_UNFIXED, "%s: no memory for %d bytes",
-		    diag, sizeof(*qhost));
+		    diag, (int)sizeof(*qhost));
 		return (GFARM_ERR_NO_MEMORY);
 	}
 	GFARM_MALLOC_ARRAY(qhost->workqs, manager->num_types);
 	if (qhost->workqs == NULL) {
 		gflog_debug(GFARM_MSG_UNFIXED, "%s: no memory for %d*%d bytes",
-		    diag, manager->num_types, sizeof(*qhost->workqs));
+		    diag, manager->num_types, (int)sizeof(*qhost->workqs));
 		return (GFARM_ERR_NO_MEMORY);
 	}
